requirements.txt
# Core Dependencies
# Python 3.8+ required (check with: python --version)
aiohttp>=3.8.0
websockets>=10.0

# Web Scraping
beautifulsoup4>=4.11.0
selenium>=4.0.0
playwright>=1.40.0
lxml>=4.9.0

# Polymarket CLOB
py-clob-client>=0.1.0
web3>=6.0.0
eth-account>=0.8.0

# Machine Learning
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.10.0

# Configuration
pyyaml>=6.0

# Logging and Monitoring
python-dotenv>=1.0.0
colorlog>=6.7.0

# Optional - for enhanced features
redis>=4.5.0  # For caching
psycopg2-binary>=2.9.0  # For PostgreSQL
pymongo>=4.3.0  # For MongoDB

# Development
pytest>=7.3.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0